+++ src/test/java/com/github/camilormz/finalreality/model/character/player/AbstractPlayerCharacterTest.java
@@ -1,15 +1,9 @@
-package com.github.camilormz.finalreality.model.character;
+package com.github.camilormz.finalreality.model.character.player;
 
-import com.github.camilormz.finalreality.model.character.player.AbstractPlayerCharacter;
-import com.github.camilormz.finalreality.model.character.player.characterclass.WhiteMage;
+import com.github.camilormz.finalreality.model.character.AbstractCharacterTest;
 import com.github.camilormz.finalreality.model.weapon.IWeapon;
-import com.github.camilormz.finalreality.model.weapon.types.*;
-import javafx.scene.input.SwipeEvent;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -18,36 +12,7 @@ import static org.junit.jupiter.api.Assertions.*;
  *
  * @author Camilo Ramírez Canales
  */
-public abstract class AbstractPlayerCharacterTest {
-
-    private final String ENEMY_NAME = "Kronos";
-    private final String AXE_NAME = "Wood Slayer";
-    private final String BOW_NAME = "The Trebuchet";
-    private final String KNIFE_NAME = "Stealthy";
-    private final String STAFF_NAME = "Shadowmaker";
-    private final String SWORD_NAME = "Infidel Redentor";
-
-    protected BlockingQueue<ICharacter> turns;
-    protected Axe testAxe;
-    protected Bow testBow;
-    protected Knife testKnife;
-    protected Staff testStaff;
-    protected Sword testSword;
-    private Enemy testEnemy;
-
-    /**
-     * Executes a setup for every test involving Playable Characters
-     */
-    @BeforeEach
-    void setUp() {
-        turns = new LinkedBlockingQueue<>();
-        testEnemy = new Enemy(ENEMY_NAME, 10, turns);
-        testAxe = new Axe(AXE_NAME, 10, 10);
-        testBow = new Bow(BOW_NAME, 10, 10);
-        testKnife = new Knife(KNIFE_NAME, 10, 10);
-        testStaff = new Staff(STAFF_NAME, 10, 10, 10);
-        testSword = new Sword(SWORD_NAME, 10, 10);
-    }
+public abstract class AbstractPlayerCharacterTest extends AbstractCharacterTest {
 
     /**
      * General test of construction for all playable characters
@@ -94,4 +59,12 @@ public abstract class AbstractPlayerCharacterTest {
      */
     @Test
     protected void subClassWeaponTest() {}
+
+    @Override
+    @Test
+    protected abstract void subClassWaitTurnTest();
+
+    @Override
+    @Test
+    protected abstract void subClassCharacterDomainTest();
 }