+++ src/test/java/com/github/camilormz/finalreality/model/character/player/characterclass/EngineerTest.java
@@ -1,8 +1,7 @@
-package com.github.camilormz.finalreality.model.character.characterclass;
+package com.github.camilormz.finalreality.model.character.player.characterclass;
 
-import com.github.camilormz.finalreality.model.character.AbstractPlayerCharacterTest;
-import com.github.camilormz.finalreality.model.character.player.characterclass.Engineer;
-import com.github.camilormz.finalreality.model.character.player.characterclass.Knight;
+import com.github.camilormz.finalreality.model.character.CharacterDomain;
+import com.github.camilormz.finalreality.model.character.player.AbstractPlayerCharacterTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -37,5 +36,18 @@ public class EngineerTest extends AbstractPlayerCharacterTest {
     @Test
     protected void subClassWeaponTest() {
         this.weaponEquipmentTest(engineer, unarmedEngineer, testBow, testStaff);
+    }
+    @Override
+    @Test
+    protected void subClassWaitTurnTest() {
+        this.waitTurnTest(engineer, 0, waitTurnTestErrorMargin);
+        engineer.equip(testBow);
+        long expectedTime = testBow.getWeight()/10;
+        this.waitTurnTest(engineer, expectedTime, waitTurnTestErrorMargin);
+    }
+    @Override
+    @Test
+    protected void subClassCharacterDomainTest() {
+        this.getCharacterDomainTest(engineer, CharacterDomain.PLAYABLE);
     }
 }